purged ext files in favor of STATIC
