+++ src/lib/maths.ts
@@ -3,13 +3,8 @@ export { default as LineSegment } from "./maths/lineSegment.js";
 export * as MathExt from "./maths/mathExt.js";
 export { default as Matrix } from "./maths/matrix.js";
 export { default as Matrix4 } from "./maths/matrix4.js";
-export * as Matrix4Ext from "./maths/matrix4Ext.js";
-export * as MatrixExt from "./maths/matrixExt.js";
 export { default as Random } from "./maths/random.js";
-export * as RandomExt from "./maths/random.js";
+export * as RandomHelper from "./maths/randomHelper.js";
 export { default as Rectangle } from "./maths/rectangle.js";
-export * as RectangleHelper from "./maths/rectangleHelper.js";
 export { default as Vector2 } from "./maths/vector2.js";
-export * as Vector2Ext from "./maths/vector2Ext.js";
 export { default as Vector3 } from "./maths/vector3.js";
-export * as Vector3Ext from "./maths/vector3Ext.js";
